requirements_clean.txt
# Core ML and Deep Learning
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Computer Vision
opencv-python==4.8.0.76
Pillow==10.0.0

# Web Framework
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
aiofiles==23.1.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
tqdm==4.65.0
python-dotenv==1.0.0

# Face Detection
facenet-pytorch==2.5.3

# Data Augmentation
albumentations==1.3.1

# Grad-CAM
grad-cam==1.4.8

# Additional utilities
requests==2.31.0





